refact(ProductReviewCommentsBlock - Comment): rename to ReviewComment
